Extract process manager test modules to separate files
