+++ test/process_managers/support/error/default_error_handling_process_manager.ex
@@ -1,0 +1,26 @@
+defmodule Commanded.ProcessManagers.DefaultErrorHandlingProcessManager do
+  @moduledoc false
+
+  alias Commanded.ProcessManagers.{
+    DefaultErrorHandlingProcessManager,
+    ExampleRouter,
+  }
+  alias Commanded.ProcessManagers.ErrorAggregate.Commands.{
+    AttemptProcess,
+  }
+  alias Commanded.ProcessManagers.ErrorAggregate.Events.{
+    ProcessStarted,
+  }
+
+  use Commanded.ProcessManagers.ProcessManager,
+    name: "DefaultErrorHandlingProcessManager",
+    router: ExampleRouter
+
+  defstruct [:process_uuid]
+
+  def interested?(%ProcessStarted{process_uuid: process_uuid}), do: {:start, process_uuid}
+
+  def handle(%DefaultErrorHandlingProcessManager{}, %ProcessStarted{process_uuid: process_uuid}) do
+    %AttemptProcess{process_uuid: process_uuid}
+  end
+end
